client/requirements.txt
playwright==1.49.1
httpx==0.28.1
websockets==14.1
pydantic==2.10.3
pyinstaller==6.11.1
pystray==0.19.5
Pillow==11.1.0
psutil==6.1.1
